requirements_v2.txt
# Core FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Dependency injection
dependency-injector>=4.41.0

# Logging
loguru>=0.7.0

# Async HTTP client
httpx>=0.25.0

# Azure services (optional - for production)
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# OpenAI (required)
openai>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Production extras
gunicorn>=21.2.0
python-jose[cryptography]>=3.3.0 